Move argument detection to top of snazzer
Checking for the existance of /etc/snazzer/exclude.patterns is achieved
through the use of sudo when snazzer is run as an unprivileged user.

By moving the argument detection to top of script we can avoid situations
where root privileges were required for commands that don't need elevated
privileges (such as --help).

Resolves: #24